requirements.txt
python-dotenv==1.0.0
Pillow==10.0.0
pytesseract==0.3.10
spacy==3.7.2
numpy==1.26.4
opencv-python==4.8.0.76
tqdm==4.66.1
setuptools>=65.5.1
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
pydantic>=1.10.0,<2.0.0
loguru==0.7.2